+++ commands/protocol_upgrade_proposal.go
@@ -18,8 +18,5 @@ func checkProtocolUpgradeProposal(cmd *commandspb.ProtocolUpgradeProposal) Error
 		errs.AddForProperty("state_variable_proposal.vega_release_tag", ErrIsRequired)
 	}
 
-	if len(cmd.DataNodeReleaseTag) == 0 {
-		errs.AddForProperty("state_variable_proposal.data_node_release_tag", ErrIsRequired)
-	}
 	return errs
 }